Graphs axis and table from seec fixed
